[AND-87] 포트폴리오 추가/수정 화면 작업 중/recyclerview, test등 필요

+++ app/src/main/java/kr/co/soogong/master/ui/requirements/done/DoneFragment.kt
@@ -79,7 +79,7 @@ class DoneFragment : BaseFragment<FragmentRequirementsDoneBinding>(
         viewModel.event.observe(viewLifecycleOwner, EventObserver { (event, value) ->
             when (event) {
                 BADGE_UPDATE -> {
-                    (parentFragment as? RequirementsBadge)?.setDoneBadge(value)
+                    (parentFragment as? RequirementsBadge)?.setDoneBadge(value as Int)
                 }
                 ASK_FOR_REVIEW_SUCCEEDED -> {
                     requireContext().toast(getString(R.string.ask_for_review_successful))
